+++ test-web/src/components/DataListView/style/excelcontrollerstyle.js
@@ -1,0 +1,20 @@
+const navy = '#0F3659';
+const style = {
+    importBtnWrapper: {
+      color: navy,
+      marginRight: '10px',
+      backgroundColor: 'white',
+      border: `1px solid ${navy}`,
+      height: '35px',
+      borderRadius: '10px',
+    },
+    exportBtnWrapper: {
+      color: navy,
+      backgroundColor: 'white',
+      border: `1px solid ${navy}`,
+      height: '35px',
+      borderRadius: '10px',
+    },
+  };
+
+  export default style
